+++ src/components/layout/sidebar/sidebar.config.ts
@@ -46,6 +46,8 @@ export const sidebarConfig: SidebarConfig = [
         children: [
           { type: 'link', label: 'sidebar.items.uiButton', href: routes.ui.buttons },
           { type: 'link', label: 'sidebar.items.uiButtonsGroup', href: routes.ui.buttonsGroup },
+          { type: 'link', label: 'sidebar.items.uiInput', href: routes.ui.inputs },
+          { type: 'link', label: 'sidebar.items.uiTextarea', href: routes.ui.textarea },
           { type: 'link', label: 'sidebar.items.toasts', href: routes.ui.toasts },
           { type: 'link', label: 'sidebar.items.emptyStates', href: routes.ui.emptyStates },
           { type: 'link', label: 'sidebar.items.errorStates', href: routes.ui.errorStates },